User Auth reveal/hide & clear forms
This commit features new user auth functionality to reveal or hide
various forms depending on the sign-in/sign-out status of a user. On
successful sign in, the sign-up/sign-in forms hide and the
change-password/sign-out forms are revealed, and the reverse happens
when the user signs out.

This commit also features the ability to clear form fields, as per
requirements. Regardless of success or error, each time a form is
submitted, all auth form fields are cleared. This feature also
required an adjustment to the index.html because the id necessary for
this to work was not on the <form> elements but were on the parent
<li> element instead.