User facing post CRUD error/success messages
This commit features user messaging for the Posts collection CRUD
actions. A majority of the changes made are in ui.js, as well as an
added div to the html document, and finally a few naming changes made
in the events.js and euth-events.js.

index.js has changes as well, but it is only to remove a refactored
function.