Bootstrapify our code, dynamically populating fields
- This commit features major changes to the layout of the html elements
in index.html and handlebars. The layout changes are in accommodation to
Bootsrap.

- This commit dynamically populates edit fields in handlebars

- This commit re-writes a lot of pre-existing SCSS elements. The reason
I did this was because everything that was being styled in SCSS was
affecting the elements I was trying to change for bootstrap. Often
times I ran into an issue that was actually the result of an SCSS
style, so I would remove those styles to make way for the bootstrap
classes and make sure they are working properly. By the end of it, I
had removed every pre-existing scss style.

NEXT STEPS: reintroduce some of the styles I removed. Now that
bootstrap is working I feel more confident targetting specific elements
to apply the styles again.