Merge conflict for previous commit
I made the previous commit, then switched to master branch, pulled the
master (recently acquired a new merge), then switched to my work
branch, rebased, handled the merge conflicts, save, and commit again,
and now we are here.

NEXT STEPS:
I noticed that the updating a post does not seem to error. We may have
to require these fields in handlebars and validate they are not empty
in events.js and ui.js. Additionally there is no user-facing error
message.