Update post success with proper behavior
This commit features the proper behavior necessary for a PATCH request.
Before, the PATCH request would go but no success or error message
would show.
This was due to an onUpdatePosts.then in events which calls upon the
entire onGetPosts function, which included a second clear message
function thus preventing any possibility of seeing a success or error
message.
By selecting specific portions of the onGetPosts event I was able to
display a success message.
Error messages are still not appearing. The update modal will close as
if the update was successful, but the modal fields do not clear, the
update does not succeed, and no error message is displayed.